+++ src/curve.rs
@@ -624,8 +624,8 @@ mod tests {
         let mut curve = Curve::new();
         curve.set_label("the-curve");
         curve.draw(x, y);
-        let b: &str = "x=np.array([1,2,3,4,5,],dtype=float)\n\
-                       y=np.array([1,4,9,16,25,],dtype=float)\n\
+        let b: &str = "x=np.array([1,2,3,4,5,])\n\
+                       y=np.array([1,4,9,16,25,])\n\
                        plt.plot(x,y,label=r'the-curve')\n";
         assert_eq!(curve.buffer, b);
         curve.clear_buffer();
@@ -640,9 +640,9 @@ mod tests {
         let mut curve = Curve::new();
         curve.set_label("the-curve");
         curve.draw_3d(x, y, z);
-        let b: &str = "x=np.array([1,2,3,4,5,],dtype=float)\n\
-                       y=np.array([1,4,9,16,25,],dtype=float)\n\
-                       z=np.array([0,0,0,1,1,],dtype=float)\n\
+        let b: &str = "x=np.array([1,2,3,4,5,])\n\
+                       y=np.array([1,4,9,16,25,])\n\
+                       z=np.array([0,0,0,1,1,])\n\
                        ax3d().plot(x,y,z,label=r'the-curve')\n";
         assert_eq!(curve.buffer, b);
     }
